requirements.txt
joblib==1.4.2
matplotlib==3.10.6
mysql_connector_repackaged==0.3.1
numpy==2.3.3
pandas==2.3.3
plotly==6.0.1
python-dotenv==1.1.1
reportlab==4.3.1
streamlit==1.44.1
streamlit_extras==0.7.8
